seperated splash from title and day comment from clock
