+++ web/src/components/clock/Clock.css
@@ -1,10 +1,3 @@
-.day {
-    font-family: Consolas, "Liberation Mono", "Courier New", monospace;
-    font-size: 30px;
-    font-weight: bold;
-    color: #d9200f;
-}
-
 .time {
     font-family: Impact, sans-serif;
     font-size: 40px;
