+++ web/src/components/dayComment/DayComment.css
@@ -1,0 +1,6 @@
+.day {
+    font-family: Consolas, "Liberation Mono", "Courier New", monospace;
+    font-size: 30px;
+    font-weight: bold;
+    color: #d9200f;
+}
